create UploadModal for upload progress
This replaces code previously in the Browse component to create a
modular, React component replacement. It deals with the abort modal and
the upload progress bar at once.
